+++ app/models/db/user.py
@@ -1,3 +1,4 @@
+from sqlalchemy import BIGINT
 from sqlalchemy.orm import mapped_column, Mapped
 
 from app.models import dto
@@ -8,7 +9,7 @@ class User(Base):
     __tablename__ = "users"
     __mapper_args__ = {"eager_defaults": True}
     id: Mapped[int] = mapped_column(primary_key=True)
-    tg_id: Mapped[int] = mapped_column(unique=True)
+    tg_id: Mapped[int] = mapped_column(BIGINT(), unique=True)
     first_name: Mapped[str]
     last_name: Mapped[str]
     username: Mapped[str]
